Security fix. Bug descriopsion from http://www.idefense.com/advisory/12.23.02.txt:
The pdftops filter in the Xpdf and CUPS packages contains an integer
overflow that can be exploited to gain the privileges of the target user
or in some cases the increased privileges of the 'lp' user if installed
setuid. There are multiple ways of exploiting this vulnerability. The
following is just one example:

A ColorSpace with 1,431,655,768 elements is created, each element having
three components. 1,431,655,768 is too large to store within a 32-bit
integer so the high bit is cut off leaving only 8 which is how much that
is actually allocated.

...
 /CS
  [
    /Indexed
    /RGB
    1431655768
    7 0 R
  ]
...

The '7 0 R' from above refers to a stream that is read into an array that
is allocated as above. The stream is read until it has reached the highest
index number, or the stream ends. If the filter supplies enough data the
application will crash when trying to access bad memory. It is possible to
exploit this condition by supplying the right length of bad memory, and
stop the stream breaking the reading. A function pointer can then be
overwritten to execute arbitrary code.

Changed files:
    xpdf-2.01-patch1 -> 1.1